:art: Style: useRating 훅으로 분리
Style: useRating 훅으로 분리
